+++ src/main/java/com/example/macaron/Store.java
@@ -2,7 +2,7 @@ package com.example.macaron;
 
 //パスワード、クーポン、自由記述が編集可能
 public class Store {
-    private final int id;
+    private final Long id;
     private final String name;
     private final String email;
     private String password;
@@ -14,7 +14,7 @@ public class Store {
     private String coupon;
     private String free_desc;
 
-    public Store(int id, String name, String email, String password, String address, int phone_number, String payment,
+    public Store(Long id, String name, String email, String password, String address, int phone_number, String payment,
             int price_range, String genre, String coupon, String free_desc) {
         this.id = id;
         this.name = name;
@@ -29,7 +29,7 @@ public class Store {
         this.free_desc = free_desc;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
