+++ src/main/java/com/example/macaron/User.java
@@ -2,20 +2,20 @@ package com.example.macaron;
 
 //メール、名前、パスワード、レビューフラグが編集可能
 public class User {
-    private final int id;
+    private final Long id;
     private String mailid;
     private String name;
     private String password;
     private boolean reviewd;
     
-    public User(int id, String mailid, String name, String password, boolean reviewd) {
+    public User(Long id, String mailid, String name, String password, boolean reviewd) {
         this.id = id;
         this.mailid = mailid;
         this.name = name;
         this.password = password;
         this.reviewd = reviewd;
     }
-    public int getId() {
+    public Long getId() {
         return id;
     }
     public String getMailid() {
